31 Jan. 2020

+++ fenrisulfur_2.0.py
@@ -171,9 +171,10 @@ async def on_message(message):
     a = await fenrir.process_commands(message)
 
     # Check if we are in dm
-    privateMessage = isinstance(message.channel, discord.abc.GuildChannel)
-    if not privateMessage:
-        if message.channel == eventsDict[hash(message.guild)].channel and message.author != fenrir.user:
+    guildMessage = isinstance(message.channel, discord.abc.GuildChannel)
+    if guildMessage \
+            and message.channel == eventsDict[hash(message.guild)].channel \
+            and message.author != fenrir.user:
             await message.delete()
 
 @fenrir.event
